requirements.txt
Flask>=2.0
PyMySQL>=1.0
Flask-Login>=0.5
cryptography>=41.0
python-dotenv>=1.0